Move subtitle-related C API functions to a separate file
